+++ quasar-utils/src/test/groovy/net/corda/plugins/quasar/QuasarPluginTest.groovy
@@ -13,14 +13,13 @@ import static org.gradle.testkit.runner.TaskOutcome.SUCCESS
 class QuasarPluginTest {
     private static final String TEST_GRADLE_USER_HOME = System.getProperty("test.gradle.user.home", ".")
     private static final String QUASAR_VERSION = QuasarPlugin.defaultVersion
-    private static final String QUASAR_CLASSIFIER = QuasarPlugin.defaultClassifier
 
     private static final String QUASAR_R3 = """\
 if (org.gradle.api.JavaVersion.current().java9Compatible) {
-    version = '0.8.0_r3'
+    version = '0.8.2_r3'
     classifier = ''
 } else {
-    version = '0.7.12_r3'
+    version = '0.7.14_r3-SNAPSHOT'
 }
 """
 
@@ -42,14 +41,14 @@ plugins {
     id 'net.corda.plugins.quasar-utils'
 }
 
-description 'Show quasar-core added to configurations'
+description 'Show quasar-core-osgi added to configurations'
     
 apply from: 'repositories.gradle'
 
 task show {
     doFirst {
         def configs = configurations.matching {
-            it.name in ['quasar', 'cordaRuntime', 'compileOnly', 'compileClasspath', 'runtimeClasspath']
+            it.name in ['quasar', 'quasarAgent', 'cordaRuntimeOnly', 'compileOnly', 'compileClasspath', 'runtimeClasspath']
         }
         configs.collectEntries { [(it.name):it.incoming.dependencies] }.each { name, dependencies ->
             dependencies.each { dep ->
@@ -62,11 +61,12 @@ task show {
 }
 """, "show"
         assertThat(output).containsOnlyOnce(
-            "quasar: co.paralleluniverse:quasar-core:jar:${QUASAR_VERSION}:".toString(),
-            "cordaRuntime: co.paralleluniverse:quasar-core:jar:${QUASAR_VERSION}:".toString(),
-            "runtimeClasspath: co.paralleluniverse:quasar-core:jar:${QUASAR_VERSION}:".toString(),
-            "compileOnly: co.paralleluniverse:quasar-core:jar:${QUASAR_VERSION}:".toString(),
-            "compileClasspath: co.paralleluniverse:quasar-core:jar:${QUASAR_VERSION}:".toString()
+            "quasar: co.paralleluniverse:quasar-core-osgi:jar:${QUASAR_VERSION}:".toString(),
+            "quasarAgent: co.paralleluniverse:quasar-core-osgi:jar:${QUASAR_VERSION}:agent".toString(),
+            "cordaRuntimeOnly: co.paralleluniverse:quasar-core-osgi:jar:${QUASAR_VERSION}:".toString(),
+            "runtimeClasspath: co.paralleluniverse:quasar-core-osgi:jar:${QUASAR_VERSION}:".toString(),
+            "compileOnly: co.paralleluniverse:quasar-core-osgi:jar:${QUASAR_VERSION}:".toString(),
+            "compileClasspath: co.paralleluniverse:quasar-core-osgi:jar:${QUASAR_VERSION}:".toString()
         )
     }
 
@@ -87,14 +87,14 @@ plugins {
     id 'net.corda.plugins.quasar-utils'
 }
 
-description 'Show quasar-core added to configurations'
+description 'Show quasar-core-osgi added to configurations'
     
 apply from: 'repositories.gradle'
 
 task show {
     doFirst {
         def configs = configurations.matching {
-            it.name in ['quasar', 'cordaRuntime', 'compileOnly', 'compileClasspath', 'runtimeClasspath']
+            it.name in ['quasar', 'quasarAgent', 'cordaRuntimeOnly', 'compileOnly', 'compileClasspath', 'runtimeClasspath']
         }
         configs.collectEntries { [(it.name):it.incoming.dependencies] }.each { name, dependencies ->
             dependencies.each { dep ->
@@ -107,11 +107,12 @@ task show {
 }
 """, "show"
         assertThat(output).containsOnlyOnce(
-            "quasar: co.paralleluniverse:quasar-core:jar:${quasarVersion}:".toString(),
-            "cordaRuntime: co.paralleluniverse:quasar-core:jar:${quasarVersion}:".toString(),
-            "runtimeClasspath: co.paralleluniverse:quasar-core:jar:${quasarVersion}:".toString(),
-            "compileOnly: co.paralleluniverse:quasar-core:jar:${quasarVersion}:".toString(),
-            "compileClasspath: co.paralleluniverse:quasar-core:jar:${quasarVersion}:".toString()
+            "quasar: co.paralleluniverse:quasar-core-osgi:jar:${quasarVersion}:".toString(),
+            "quasarAgent: co.paralleluniverse:quasar-core-osgi:jar:${quasarVersion}:agent".toString(),
+            "cordaRuntimeOnly: co.paralleluniverse:quasar-core-osgi:jar:${quasarVersion}:".toString(),
+            "runtimeClasspath: co.paralleluniverse:quasar-core-osgi:jar:${quasarVersion}:".toString(),
+            "compileOnly: co.paralleluniverse:quasar-core-osgi:jar:${quasarVersion}:".toString(),
+            "compileClasspath: co.paralleluniverse:quasar-core-osgi:jar:${quasarVersion}:".toString()
         )
     }
 
@@ -131,7 +132,6 @@ plugins {
     id 'net.corda.plugins.quasar-utils'
 }
 
-apply plugin: 'net.corda.plugins.quasar-utils'
 apply from: 'repositories.gradle'
 
 quasar {
@@ -141,7 +141,7 @@ quasar {
 task show {
     doFirst {
         def configs = configurations.matching {
-            it.name in ['quasar', 'cordaRuntime', 'compileOnly', 'compileClasspath', 'runtimeClasspath']
+            it.name in ['quasar', 'quasarAgent', 'cordaRuntimeOnly', 'compileOnly', 'compileClasspath', 'runtimeClasspath']
         }
         configs.collectEntries { [(it.name):it.incoming.dependencies] }.each { name, dependencies ->
             dependencies.each { dep ->
@@ -154,55 +154,12 @@ task show {
 }
 """, "show"
         assertThat(output).containsOnlyOnce(
-            "quasar: co.paralleluniverse:quasar-core:jar:${quasarVersion}:".toString(),
-            "cordaRuntime: co.paralleluniverse:quasar-core:jar:${quasarVersion}:".toString(),
-            "runtimeClasspath: co.paralleluniverse:quasar-core:jar:${quasarVersion}:".toString(),
-            "compileOnly: co.paralleluniverse:quasar-core:jar:${quasarVersion}:".toString(),
-            "compileClasspath: co.paralleluniverse:quasar-core:jar:${quasarVersion}:".toString()
-        )
-    }
-
-    @Test
-    void checkLocalOverriddenClassifierVersionIsUsed() {
-        def quasarVersion = '0.7.12_r3'
-        def quasarClassifier = 'jdk8'
-        assertThat(quasarVersion).isNotEqualTo(QUASAR_VERSION)
-        assertThat(quasarClassifier).isNotEqualTo(QUASAR_CLASSIFIER)
-
-        def output = runGradleFor """
-plugins {
-    id 'net.corda.plugins.quasar-utils'
-}
-
-apply from: 'repositories.gradle'
-
-quasar {
-    group = 'co.paralleluniverse'
-    version = '${quasarVersion}'
-    classifier = '${quasarClassifier}'
-}
-
-task show {
-    doFirst {
-        def configs = configurations.matching {
-            it.name in ['quasar', 'cordaRuntime', 'compileOnly', 'compileClasspath', 'runtimeClasspath']
-        }
-        configs.collectEntries { [(it.name):it.incoming.dependencies] }.each { name, dependencies ->
-            dependencies.each { dep ->
-                dep.artifacts.each { art ->
-                    println "\$name: \${dep.group}:\${dep.name}:\${art.extension}:\${dep.version}:\${art.classifier ?: ''}"
-                }
-            }
-        }
-    }
-}
-""", "show"
-        assertThat(output).containsOnlyOnce(
-            "quasar: co.paralleluniverse:quasar-core:jar:${quasarVersion}:${quasarClassifier}".toString(),
-            "cordaRuntime: co.paralleluniverse:quasar-core:jar:${quasarVersion}:${quasarClassifier}".toString(),
-            "runtimeClasspath: co.paralleluniverse:quasar-core:jar:${quasarVersion}:${quasarClassifier}".toString(),
-            "compileOnly: co.paralleluniverse:quasar-core:jar:${quasarVersion}:${quasarClassifier}".toString(),
-            "compileClasspath: co.paralleluniverse:quasar-core:jar:${quasarVersion}:${quasarClassifier}".toString()
+            "quasar: co.paralleluniverse:quasar-core-osgi:jar:${quasarVersion}:".toString(),
+            "quasarAgent: co.paralleluniverse:quasar-core-osgi:jar:${quasarVersion}:agent".toString(),
+            "cordaRuntimeOnly: co.paralleluniverse:quasar-core-osgi:jar:${quasarVersion}:".toString(),
+            "runtimeClasspath: co.paralleluniverse:quasar-core-osgi:jar:${quasarVersion}:".toString(),
+            "compileOnly: co.paralleluniverse:quasar-core-osgi:jar:${quasarVersion}:".toString(),
+            "compileClasspath: co.paralleluniverse:quasar-core-osgi:jar:${quasarVersion}:".toString()
         )
     }
 
@@ -213,7 +170,7 @@ plugins {
     id 'net.corda.plugins.quasar-utils'
 }
 
-description 'Show quasar-core added to configurations'
+description 'Show quasar-core-osgi added to configurations'
     
 apply from: 'repositories.gradle'
 
@@ -221,7 +178,7 @@ apply plugin: 'net.corda.plugins.quasar-utils'
 
 task show {
     doFirst {
-        def configs = configurations.matching { it.name in ['quasar', 'cordaRuntime', 'compileClasspath', 'compileOnly', 'runtimeClasspath'] }
+        def configs = configurations.matching { it.name in ['quasar', 'quasarAgent', 'cordaRuntimeOnly', 'compileClasspath', 'compileOnly', 'runtimeClasspath'] }
         configs.collectEntries { [(it.name):it] }.each { name, files ->
             files.each { file ->
                 println "\$name: \${file.name}"
@@ -231,7 +188,8 @@ task show {
 }
 """, "show"
         assertThat(output.findAll { it.startsWith("quasar:") }).hasSize(1)
-        assertThat(output.findAll { it.startsWith("cordaRuntime:") }).hasSize(1)
+        assertThat(output.findAll { it.startsWith("quasarAgent:") }).hasSize(1)
+        assertThat(output.findAll { it.startsWith("cordaRuntimeOnly:") }).hasSize(1)
         assertThat(output.findAll { it.startsWith("compileOnly:") }).hasSize(1)
         assertThat(output.findAll { it.startsWith("compileClasspath:") }).hasSize(1)
         assertThat(output.findAll { it.startsWith("runtimeClasspath:") }.size()).isGreaterThan(1)
@@ -244,7 +202,7 @@ plugins {
     id 'net.corda.plugins.quasar-utils'
 }
 
-description 'Show quasar-core added to test JVM arguments'
+description 'Show quasar-core-osgi added to test JVM arguments'
 
 apply from: 'repositories.gradle'
 
@@ -269,7 +227,7 @@ test {
 }
 """, "test"
         assertThat(output).anyMatch {
-            it.startsWith("TEST-JVM: -javaagent:") && it.contains("quasar-core-") && it.endsWith(".jar")
+            it.startsWith("TEST-JVM: -javaagent:") && it.contains("quasar-core-osgi-") && it.endsWith(".jar")
         }.anyMatch {
             it == "TEST-JVM: -Dco.paralleluniverse.fibers.verifyInstrumentation"
         }
@@ -288,7 +246,7 @@ plugins {
     id 'net.corda.plugins.quasar-utils'
 }
 
-description 'Show quasar-core added to test JVM arguments'
+description 'Show quasar-core-osgi added to test JVM arguments'
 
 apply from: 'repositories.gradle'
 
@@ -325,7 +283,7 @@ plugins {
     id 'net.corda.plugins.quasar-utils'
 }
 
-description 'Show quasar-core added to test JVM arguments'
+description 'Show quasar-core-osgi added to test JVM arguments'
 
 apply from: 'repositories.gradle'
 
@@ -362,7 +320,7 @@ plugins {
     id 'net.corda.plugins.quasar-utils'
 }
 
-description 'Show quasar-core added to test JVM arguments'
+description 'Show quasar-core-osgi added to test JVM arguments'
 
 apply from: 'repositories.gradle'
 
@@ -408,7 +366,7 @@ plugins {
     id 'net.corda.plugins.quasar-utils'
 }
 
-description 'Show quasar-core added to test JVM arguments'
+description 'Show quasar-core-osgi added to test JVM arguments'
 
 apply from: 'repositories.gradle'
 """, "test").buildAndFail()