+++ changelog.md
@@ -5,6 +5,7 @@
 ### Version 6.0.0
 
 * Compatible with Gradle 6.x.
+* `quasar-utils`: Upgrade to use Quasar 0.8.2_r3 OSGi bundles.
 
 ## Version 5